refactor(ale_claw): Tier 4 — split replay pipeline out of session.py
Moves the transcript-replay pipeline (13 functions + the two session-internal
regexes _THINKING_BLOCK_RE / _BASE64_IMAGE_RE) to replay.py. On inspection the
block had zero runtime coupling to core session symbols — it touches
TranscriptEntry only as an annotation and the regexes are used nowhere else
(context.py has its own copy) — so the same cycle-free pattern applies:
`from __future__ import annotations` + TYPE_CHECKING import of TranscriptEntry,
and replay imports _normalize_actions from canonical (which never imports
session). session.py re-exports the four public entry points
(build_replay_messages / sanitize_history / limit_history_turns /
convert_to_responses_api_items) so all `from .session import ...` paths
(__init__/agent_loop/tests) are unchanged; dropped its now-unused `import re`
and the canonical _normalize_actions import. session.py 1097 -> 517 LOC.
Pure move, no behavior change; all paths resolve to identical objects;
full suite green (928 passed, 3 skipped).
